+++ src/providers/AuthProvider.jsx
@@ -2,7 +2,11 @@ import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useNavigate, useLocation } from "react-router-dom";
 import { jwtDecode } from "jwt-decode";
-
+import {
+  getProductCategories,
+  getProductStaticValuesByName,
+} from "../redux/actions/productActions";
+import { setProductCategories } from "../redux";
 // import { loginUser, setUserInfoError, logoutUser } from "../redux";
 import {
   axiosWithCredentials,
@@ -20,7 +24,53 @@ export function AuthProvider({ children }) {
   //   location.pathname !== "/" && !publicRoutes.includes(location.pathname);
 
   const [loading, setLoading] = useState(true);
+  const fetchCategories = async () => {
+    try {
+      // fetch dynamic categories
+      const dynamicCategories = await getProductCategories();
 
+      // fetch unqiue values for thickness, width, and weight from all products list
+      const staticValuesPromises = [
+        getProductStaticValuesByName("thickness"),
+        getProductStaticValuesByName("width"),
+        // getProductStaticValuesByName("weight"),
+      ];
+      // backend not sending weight unique values :) need to add from backend
+      const [thicknessRes, widthRes, weightRes] = await Promise.all(
+        staticValuesPromises
+      );
+
+      // transform unqiue values into the same format as categories
+      const transformUniqiueValues = (name, values) => ({
+        id: Math.floor(Math.random() * 1000),
+        name,
+        choices: values?.map((value) => ({
+          id: Number(value),
+          category: name,
+          name_en: value,
+          name_nl: value,
+        })),
+      });
+
+      const uniqueCategories = [
+        transformUniqiueValues("thickness", thicknessRes.values),
+        transformUniqiueValues("width", widthRes.values),
+        // transformUniqiueValues("weight", weightRes.values),
+      ];
+
+
+      // combine dynamic and unique static categories
+      const combinedCategories = [...dynamicCategories, ...uniqueCategories];
+      dispatch(setProductCategories(combinedCategories));
+    } catch (error) {
+      console.error("Error fetching categories:", error);
+      dispatch(setProductCategories([]));
+    }
+  };
+
+  useEffect(() => {
+    fetchCategories();
+  }, []);
   // const handleLogoutUser = async () => {
   //   setAccessToken("");
   //   dispatch(logoutUser());
